minimal wasm.. But I don't think it works.
