+++ Cargo.toml
@@ -1,9 +1,4 @@
 [package]
-name = "rust_calc"
+name = "rustcalc"
 version = "0.1.0"
 edition = "2021"
-
-[dependencies]
-
-[lib]
-crate-type = ["cdylib"]
